Refactor actions
Instead of requiring that each implementation of `__getitem__` first call
`HasActions.try_action`, instead the HasActionsMeta wraps the `__getitem__`
function call.

This means that it's now not possible to accidently leave out the
`try_action`, which fixes a few bugs.

However, that means it's also now not possible to override the action
behaviour.